move constants over to config, where it belongs
