+++ app/lib/features/chat/widgets/emoji/emoji_row.dart
@@ -21,7 +21,7 @@
  */
 
 import 'package:acter/common/themes/app_theme.dart';
-import 'package:acter/common/utils/constants.dart';
+import 'package:acter/config/constants.dart';
 import 'package:acter/common/widgets/emoji_picker_widget.dart';
 import 'package:atlas_icons/atlas_icons.dart';
 import 'package:flutter/material.dart';
