+++ app/lib/router/router.dart
@@ -1,7 +1,7 @@
 import 'package:acter/common/extensions/acter_build_context.dart';
 import 'package:acter/common/extensions/options.dart';
 import 'package:acter/common/pages/not_found.dart';
-import 'package:acter/common/utils/constants.dart';
+import 'package:acter/config/constants.dart';
 import 'package:acter/common/utils/routes.dart';
 import 'package:acter/features/main/app_shell.dart';
 import 'package:acter/features/home/providers/client_providers.dart';
