+++ app/test/features/auth/pages/login_page_test.dart
@@ -1,4 +1,4 @@
-import 'package:acter/common/utils/constants.dart';
+import 'package:acter/config/constants.dart';
 import 'package:acter/features/auth/pages/login_page.dart';
 import 'package:acter/l10n/generated/l10n.dart';
 import 'package:flutter/material.dart';
